pyproject.toml
[project]
name = "coinbaseparse"
version = "0.1.0"
description = "Парсер coinbase"
authors = [
    {name = "linsen00777@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "ruff (>=0.14.6,<0.15.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "httpx (==0.23.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "apscheduler (>=3.11.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
